Rename linalg -> matrices
